+++ src/main/java/net/t7seven7t/craftfx/effect/BoundArmor.java
@@ -1,0 +1,15 @@
+/**
+ * Copyright (C) 2013 t7seven7t
+ */
+package net.t7seven7t.craftfx.effect;
+
+/**
+ * @author t7seven7t
+ */
+public class BoundArmor {
+
+	/**
+	 * TODO: Event listening of some sort so that effect can manage when player removes armor 
+	 */
+	
+}