+++ src/main/java/net/t7seven7t/craftfx/effect/Explosion.java
@@ -3,32 +3,33 @@
  */
 package net.t7seven7t.craftfx.effect;
 
-import net.t7seven7t.craftfx.Trigger;
 import org.bukkit.Location;
-import org.bukkit.configuration.ConfigurationSection;
-import org.bukkit.inventory.ItemStack;
 
 /**
  * @author t7seven7t
  */
 public class Explosion extends Effect {
 
+	private static final String POWER_PATH = "power";
+	private static final String SET_FIRE_PATH = "set-fire";
+	private static final String BREAK_BLOCKS_PATH = "break-blocks";
+	
 	float power = 1.0f;
 	boolean setFire = false;
 	boolean breakBlocks = false;
 	
-	public Explosion(Trigger trigger, ItemStack item, ConfigurationSection config) {
-		super(trigger, item);
+	@Override
+	public void initialize() {
 		
-		if (config.contains("power"))
-			power = (float) config.getDouble("power");
+		if (getConfig().isDouble(POWER_PATH))
+			power = (float) getConfig().getDouble(POWER_PATH);
 		
-		if (config.contains("set-fire"))
-			setFire = config.getBoolean("set-fire");
+		if (getConfig().isBoolean(SET_FIRE_PATH))
+			setFire = getConfig().getBoolean(SET_FIRE_PATH);
 		
-		if (config.contains("break-blocks"))
-			breakBlocks = config.getBoolean("break-blocks");
-		
+		if (getConfig().isBoolean(BREAK_BLOCKS_PATH))
+			breakBlocks = getConfig().getBoolean(BREAK_BLOCKS_PATH);
+
 	}
 
 	@Override