+++ src/main/java/net/t7seven7t/craftfx/effect/Invisibility.java
@@ -1,0 +1,13 @@
+/**
+ * Copyright (C) 2013 t7seven7t
+ */
+package net.t7seven7t.craftfx.effect;
+
+/**
+ * @author t7seven7t
+ */
+public class Invisibility extends Effect {
+	
+	
+	
+}