+++ src/main/java/net/t7seven7t/craftfx/effect/Teleport.java
@@ -1,0 +1,15 @@
+/**
+ * Copyright (C) 2013 t7seven7t
+ */
+package net.t7seven7t.craftfx.effect;
+
+/**
+ * @author t7seven7t
+ */
+public class Teleport extends Effect {
+
+	/**
+	 * TODO: More run types so that effect knows player + location.
+	 */
+	
+}